chore: rename user-skill in kebab-case
